+++ examples/import_export.py
@@ -1,7 +1,7 @@
 from surrealist import Surreal
 
 # we have to use http transport for import/export
-surreal = Surreal("http://127.0.0.1:8000", "test", "test", credentials=("root", "root"), use_http=True)
+surreal = Surreal("http://127.0.0.1:8000", "test", "test", credentials=("user_db", "user_db"), use_http=True)
 with surreal.connect() as connection:
     # first, we export data from SurrealDB
     with open("exported.surql", "wt", encoding="UTF-8") as file: