+++ examples/surreal_ql/ql_insert_examples.py
@@ -5,7 +5,7 @@ from surrealist import Database
 
 # Notice: all queries below not executed, just generate representation.
 # To run it against SurrealDB, you need to use run method
-with Database("http://127.0.0.1:8000", 'test', 'test', ('root', 'root'), use_http=True) as db:
+with Database("http://127.0.0.1:8000", 'test', 'test', credentials=("root", "root")) as db:
     data = {
         'name': 'SurrealDB',
         'founded': "2021-09-10",