+++ src/surrealist/connections/connection.py
@@ -346,6 +346,21 @@ class Connection(ABC):
     @abstractmethod
     def update(self, table_name: str, data: Dict, record_id: Optional[str] = None) -> SurrealResult:
         """
+        This method can be used to update record in the database if it exists.
+        So all old fields will be deleted, and new will be added, if you wand just to add field to record,
+        keeping old ones -use **merge** method instead.
+        If a record with specified id does not exist, it will NOT be created, and no errors returns back,
+        just empty result
+
+        Note: if you want to create/replace one record, you should specify recordID in table_name or in record_id, but
+        not in data parameters.
+
+        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/update
+        """
+
+    @abstractmethod
+    def upsert(self, table_name: str, data: Dict, record_id: Optional[str] = None) -> SurrealResult:
+        """
         This method can be used to update or create record in the database. So all old fields will be deleted, and new
         will be added, if you wand just to add field to record, keeping old ones -use **merge** method instead. If
         a record with specified id does not exist, it will be created, if exist - all fields will be replaced
@@ -353,7 +368,7 @@ class Connection(ABC):
         Note: if you want to create/replace one record, you should specify recordID in table_name or in record_id, but
         not in data parameters.
 
-        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/update
+        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/upsert
         """
 
     @abstractmethod