+++ src/surrealist/connections/ws_connection.py
@@ -400,8 +400,8 @@ class WebSocketConnection(Connection):
     def update(self, table_name: str, data: Dict, record_id: Optional[str] = None) -> SurrealResult:
         """
         This method can be used to update or modify records in the database. So all old fields will be deleted, and new
-        will be added, if you wand just to add field to record, keeping old ones -use **merge** method instead.If
-        record with specified id does not exist it will be created, if it exists - all fields will be replaced
+        will be added, if you wand just to add field to record, keeping old ones -use **merge** method instead.
+        If a record with specified id does not exist, it will NOT be created, use **upsert** for that.
 
         Note: if you want to create/replace one record, you should specify recordID in table_name or in record_id, but
         not in data parameters.
@@ -426,6 +426,36 @@ class WebSocketConnection(Connection):
         table_name = table_name if record_id is None else f"{table_name}:{record_id}"
         _data = {"method": "update", "params": [table_name, data]}
         logger.info("Operation: UPDATE. Path: %s, data: %s", crop_data(table_name), crop_data(str(data)))
+        return self._run(_data)
+
+    @connected
+    def upsert(self, table_name: str, data: Dict, record_id: Optional[str] = None) -> SurrealResult:
+        """
+        This method can be used to create or update records in the database. So all old fields will be deleted, and new
+        will be added, if you wand just to add field to record, keeping old ones -use **merge** method instead.
+        If a record with specified id does not exist, it will be created, if it exists - all fields will be replaced
+
+        Note: if you want to create/replace one record, you should specify recordID in table_name or in record_id, but
+        not in data parameters.
+
+        Refer to: https://docs.surrealdb.com/docs/integration/websocket#upsert
+
+        Refer to: https://docs.surrealdb.com/docs/surrealql/statements/upsert
+
+        Example:
+        websocket_connection.upsert("person:my_id", {"name": "Alex Doe"}) # record with specified id will be now
+        {"name": "Alex Doe"}, all data stored in record before will be deleted
+
+        Notice: do not specify id twice, for example, in table name and in record_id, it will cause error
+
+        :param table_name: table name or table name with record_id to upsert
+        :param data: dict with data to create
+        :param record_id: optional parameter, if it exists it will transform table_name to "table_name:record_id"
+        :return: result of request
+        """
+        table_name = table_name if record_id is None else f"{table_name}:{record_id}"
+        _data = {"method": "upsert", "params": [table_name, data]}
+        logger.info("Operation: UPSERT. Path: %s, data: %s", crop_data(table_name), crop_data(str(data)))
         return self._run(_data)
 
     @connected