+++ tests/integration_tests/test_ql_db.py
@@ -2,7 +2,7 @@ import time
 from typing import List
 from unittest import TestCase, main
 
-from surrealist import Database, WrongCallError
+from surrealist import Database, WrongCallError, Surreal, SurrealConnectionError
 from tests.integration_tests.utils import URL, get_random_series
 
 
@@ -45,6 +45,37 @@ class TestDatabase(TestCase):
         with Database(URL, 'test', 'test', credentials=('user_db', 'user_db'), use_http=True) as db:
             with self.assertRaises(WrongCallError):
                 db.live(1, 2)
+
+    def test_inactive_connections_fails(self):
+        surreal = Surreal(URL, 'test', 'test', credentials=('user_db', 'user_db'))
+        connection = surreal.connect()
+        connection.close()
+        with self.assertRaises(SurrealConnectionError) as e:
+            Database.from_connection(connection)
+        self.assertEqual(str(e.exception), "Cant use connection which is not active")
+
+    def test_root_level_fails(self):
+        surreal = Surreal(URL, credentials=('root', 'root'))
+        connection = surreal.connect()
+        with self.assertRaises(SurrealConnectionError) as e:
+            Database.from_connection(connection)
+        self.assertEqual(str(e.exception), "Can use only database level connections, specify namespace and database")
+
+    def test_ns_level_fails(self):
+        surreal = Surreal(URL, namespace="test", credentials=('user_ns', 'user_ns'))
+        connection = surreal.connect()
+        with self.assertRaises(SurrealConnectionError) as e:
+            Database.from_connection(connection)
+        self.assertEqual(str(e.exception), "Can use only database level connections, specify namespace and database")
+
+    def test_from_connections_works(self):
+        with Surreal(URL, credentials=('root', 'root')).connect() as connection:
+            connection.use("test", "test")
+            db = Database.from_connection(connection)
+            self.assertEqual("test", db._namespace)
+            self.assertEqual("test", db._database)
+            self.assertEqual(connection, db._connection)
+
 
 
 if __name__ == '__main__':